EuropeSocialEconomicsDashboard/requirements.txt
dash
pandas
numpy
plotly
flask-caching
gunicorn